+++ src/pyp/detect/joint.py
@@ -8,6 +8,7 @@ import time
 import numpy as np
 from pathlib import Path
 
+from pyp.inout.image.core import get_image_dimensions
 from pyp.inout.metadata import pyp_metadata
 from pyp.system import local_run, project_params
 from pyp.system import local_run, mpi
@@ -22,8 +23,11 @@ NN_INIT_COMMANDS_2D = f"{NN_INIT_COMMANDS}; export PYTHONPATH=$PYTHONPATH:$PYP_D
 NN_INIT_COMMANDS_3D = f"{NN_INIT_COMMANDS}; export PYTHONPATH=$PYTHONPATH:$PYP_DIR/external/cet_pick; micromamba run -n milopyp"
 
 def bin_image( input, output, binning):
-    command = f"{get_imod_path()}/bin/newstack '{input}' '{output}' -bin {binning}; rm -f '{output}~'"
-    local_run.run_shell_command(command)
+    if binning > 1:
+        command = f"{get_imod_path()}/bin/newstack '{input}' '{output}' -shrink {binning}; rm -f '{output}~'"
+        local_run.run_shell_command(command)
+    else:
+        shutil.copy2(input,output)
 
 def sprtrain(args):
 
@@ -651,9 +655,14 @@ def milotrain(args):
             if not os.path.exists(os.path.join("mrc", file + "_bin.ali")):
                 try:
                     binning = args["tomo_rec_binning"]
-                    comm = "{0}/bin/newstack -input mrc/{1}.mrc -output mrc/{1}_bin.ali -mode 2 -origin -linear -bin {2}".format( get_imod_path(), file, binning )
-
-                    [output, error] =local_run.run_shell_command(comm)
+                    imod_binning_option = f"-shrink {binning}" if binning > 1 else ""
+                    x, y, _ = get_image_dimensions(f'mrc/{file}.mrc')
+                    size_x = round(x / binning)
+                    size_x -= size_x % 2
+                    size_y = round(y / binning)
+                    size_y -= size_y % 2
+                    comm = "{0}/bin/newstack -input mrc/{1}.mrc -output mrc/{1}_bin.ali -mode 2 -origin -linear {2} -size {3},{4}".format( get_imod_path(), file, imod_binning_option, size_x, size_y )
+                    [output, _] = local_run.run_shell_command(comm)
                 except:
                     raise Exception("Can't find aligned tilt series images")
 
@@ -918,9 +927,14 @@ def miloeval(args):
                     if not os.path.exists(os.path.join("mrc", file + "_bin.ali")):
                         try:
                             binning = args["tomo_rec_binning"]
-                            comm = "{0}/bin/newstack -input mrc/{1}.mrc -output mrc/{1}_bin.ali -mode 2 -origin -linear -bin {2}".format( get_imod_path(), file, binning )
-
-                            [output, error] =local_run.run_shell_command(comm)
+                            imod_binning_option = f"-shrink {binning}" if binning > 1 else ""
+                            x, y, _ = get_image_dimensions(f'mrc/{file}.mrc')
+                            size_x = round(x / binning)
+                            size_x -= size_x % 2
+                            size_y = round(y / binning)
+                            size_y -= size_y % 2
+                            comm = "{0}/bin/newstack -input mrc/{1}.mrc -output mrc/{1}_bin.ali -mode 2 -origin -linear {2} -size {3},{4}".format( get_imod_path(), file, imod_binning_option, size_x, size_y )
+                            local_run.run_shell_command(comm)
                         except:
                             raise Exception("Can't find aligned tilt series images")